renamed shared folder for consistency
