+++ client/src/components/Login/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import Box from '@mui/material/Box'
 import { Tab, Tabs } from '@mui/material'
 import BusinessLogin from './BusinessLogin'
-import TabPanel from '../shared/TabPanel'
+import TabPanel from '../Shared/TabPanel'
 import Container from '@mui/material/Container'
 import { AuthProps } from '../../utils/sharedProps'
 import InfluencerLogin from './InfluencerLogin'
