+++ client/src/components/Nav/index.tsx
@@ -5,7 +5,7 @@ import Box from '@mui/material/Box'
 import MenuIcon from '@mui/icons-material/Menu'
 import * as React from 'react'
 import Button from '@mui/material/Button'
-import { StyledTitle } from '../shared/StyledTitle'
+import { StyledTitle } from '../Shared/StyledTitle'
 import UserMenu from './UserMenu'
 import { AuthProps } from '../../utils/sharedProps'
 import { getNavItems } from '../../utils/paths'
