+++ client/src/components/Profile/BusinessProfile.tsx
@@ -3,6 +3,8 @@ import Box from '@mui/material/Box'
 
 export default function BusinessProfile (): JSX.Element {
   return (
-    <Box></Box>
+    <Box>
+
+    </Box>
   )
 }
