+++ client/src/components/Signup/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import Box from '@mui/material/Box'
 import { Tab, Tabs } from '@mui/material'
 import BusinessSignup from './BusinessSignup'
-import TabPanel from '../shared/TabPanel'
+import TabPanel from '../Shared/TabPanel'
 import Container from '@mui/material/Container'
 import { AuthProps } from '../../utils/sharedProps'
 import InfluencerSignup from './InfluencerSignup'
